Adds alert banner
